Add Moveable class
*Add Moveable abstract class
*Refactor NPCController to use Moveable
*Refactor PlayerController to use Moveable
