+++ Assets/Ravar/Scripts/Character/Moveable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e8ee05c50daa30943a223bf9e2911cab
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
